+++ src/crm_stdio_port.c
@@ -92,21 +92,21 @@ int is_stdout_err_or_null(FILE *f)
 
 
 
-int fwrite4stdio(const char *str, size_t len, FILE *out)
+ssize_t fwrite4stdio(const char *str, size_t len, FILE *out)
 {
-    int ret = 0;
+    ssize_t ret = 0;
 
 #if !defined (BUFSIZ)
 #define BUFSIZ4FWRITE           256
 #else
 #define BUFSIZ4FWRITE           (BUFSIZ / 2)
 #endif
-    int count_written = 0;
+    ssize_t count_written = 0;
 
     // Windows does NOT like it when you try to write large pieces of text to stdout all at once, so we do it in bits and pieces.
     for ( ; len > 0;)
     {
-        ret = (int)fwrite(str, 1, CRM_MIN(BUFSIZ4FWRITE, len), out);
+        ret = (ssize_t)fwrite(str, 1, CRM_MIN(BUFSIZ4FWRITE, len), out);
         if (ret > 0)
         {
             count_written += ret;
